// resources/js/api/generated/models/securityAuditLogInfo.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface SecurityAuditLogInfo {
  actor_ip: string;
  actor_user_agent: string;
  /**
   * @minimum 0
   * @nullable
   */
  actor_user_id: number | null;
  actor_username: string;
  created_at: string;
  event_category: string;
  event_type: string;
  failure_reason: string;
  /** @minimum 0 */
  id: number;
  metadata: string;
  /**
   * @minimum 0
   * @nullable
   */
  server_id: number | null;
  session_id: string;
  severity: string;
  stack_name: string;
  success: boolean;
  /**
   * @minimum 0
   * @nullable
   */
  target_id: number | null;
  target_name: string;
  target_type: string;
  /**
   * @minimum 0
   * @nullable
   */
  target_user_id: number | null;
  updated_at: string;
}